project-artha/requirements.txt
google-adk>=0.1.0
google-genai>=0.8.0
aiohttp>=3.8.0
sqlalchemy>=1.4.0
asyncio-mqtt>=0.11.0
pandas>=1.5.0
numpy>=1.24.0
deprecated
firebase-admin>=6.5.0
python-dotenv
google-cloud-aiplatform[agent_engines,adk]>=1.60.0